# projects/content-factory/packages/contracts/openapi/content-factory.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Content Factory API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/content-items": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Create Content Item",
        "operationId": "create_content_item_api_content_items_post",
        "parameters": [
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentItemCreateRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "content"
        ],
        "summary": "List Content Items",
        "operationId": "list_content_items_api_content_items_get",
        "parameters": [
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentItemListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/content-items/{content_item_id}": {
      "get": {
        "tags": [
          "content"
        ],
        "summary": "Get Content Item",
        "operationId": "get_content_item_api_content_items__content_item_id__get",
        "parameters": [
          {
            "name": "content_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Item Id"
            }
          },
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/content-items/{content_item_id}/plan": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Plan Content Item",
        "operationId": "plan_content_item_api_content_items__content_item_id__plan_post",
        "parameters": [
          {
            "name": "content_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Item Id"
            }
          },
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ContentPlanRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ContentItemRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/content-items/{content_item_id}/submit-review": {
      "post": {
        "tags": [
          "content"
        ],
        "summary": "Submit Content Item For Review",
        "operationId": "submit_content_item_for_review_api_content_items__content_item_id__submit_review_post",
        "parameters": [
          {
            "name": "content_item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Content Item Id"
            }
          },
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/tasks": {
      "get": {
        "tags": [
          "review"
        ],
        "summary": "List Review Tasks",
        "operationId": "list_review_tasks_api_review_tasks_get",
        "parameters": [
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/tasks/{task_id}/approve": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Approve Review Task",
        "operationId": "approve_review_task_api_review_tasks__task_id__approve_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/review/tasks/{task_id}/request-rework": {
      "post": {
        "tags": [
          "review"
        ],
        "summary": "Request Rework",
        "operationId": "request_rework_api_review_tasks__task_id__request_rework_post",
        "parameters": [
          {
            "name": "task_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Task Id"
            }
          },
          {
            "name": "cf_session",
            "in": "cookie",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Cf Session"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDecisionRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewTaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ContentChannel": {
        "type": "string",
        "enum": [
          "instagram_reels",
          "youtube_shorts"
        ],
        "title": "ContentChannel"
      },
      "ContentItemCreateRequest": {
        "properties": {
          "brand_id": {
            "type": "string",
            "title": "Brand Id"
          },
          "avatar_id": {
            "type": "string",
            "title": "Avatar Id"
          },
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Title"
          },
          "script": {
            "type": "string",
            "maxLength": 20000,
            "minLength": 1,
            "title": "Script"
          },
          "channel": {
            "$ref": "#/components/schemas/ContentChannel"
          }
        },
        "type": "object",
        "required": [
          "brand_id",
          "avatar_id",
          "title",
          "script",
          "channel"
        ],
        "title": "ContentItemCreateRequest"
      },
      "ContentItemListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ContentItemRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ContentItemListResponse"
      },
      "ContentItemRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "brand_id": {
            "type": "string",
            "title": "Brand Id"
          },
          "avatar_id": {
            "type": "string",
            "title": "Avatar Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "script": {
            "type": "string",
            "title": "Script"
          },
          "channel": {
            "$ref": "#/components/schemas/ContentChannel"
          },
          "status": {
            "$ref": "#/components/schemas/ContentStatus"
          },
          "planned_publish_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Publish At"
          },
          "created_by_user_id": {
            "type": "string",
            "title": "Created By User Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "brand_id",
          "avatar_id",
          "title",
          "script",
          "channel",
          "status",
          "planned_publish_at",
          "created_by_user_id",
          "created_at",
          "updated_at"
        ],
        "title": "ContentItemRead"
      },
      "ContentPlanRequest": {
        "properties": {
          "planned_publish_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Planned Publish At"
          }
        },
        "type": "object",
        "title": "ContentPlanRequest"
      },
      "ContentStatus": {
        "type": "string",
        "enum": [
          "draft",
          "planned",
          "review",
          "approved",
          "rework"
        ],
        "title": "ContentStatus"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ReviewDecisionRequest": {
        "properties": {
          "decision_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10000
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Notes"
          }
        },
        "type": "object",
        "title": "ReviewDecisionRequest"
      },
      "ReviewTaskListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ReviewTaskRead"
            },
            "type": "array",
            "title": "Items"
          }
        },
        "type": "object",
        "required": [
          "items"
        ],
        "title": "ReviewTaskListResponse"
      },
      "ReviewTaskRead": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "content_item_id": {
            "type": "string",
            "title": "Content Item Id"
          },
          "assigned_to_user_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assigned To User Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewTaskStatus"
          },
          "decision_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Decision Notes"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "content_item_id",
          "assigned_to_user_id",
          "status",
          "decision_notes",
          "completed_at",
          "created_at",
          "updated_at"
        ],
        "title": "ReviewTaskRead"
      },
      "ReviewTaskStatus": {
        "type": "string",
        "enum": [
          "open",
          "approved",
          "rework",
          "cancelled"
        ],
        "title": "ReviewTaskStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}